refactor everything possible into AdoQueryBase.
