+++ Highway/src/Highway.Data.EntityFramework/QueryObjects/AdoCommandBase.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Data;
+
+namespace Highway.Data
+{
+    public abstract class AdoCommandBase : AdvancedCommand
+    {
+        protected abstract IEnumerable<IDataParameter> Parameters { get; }
+    }
+}
