+++ Highway/src/Highway.Data.EntityFramework/QueryObjects/AdoScalarBase.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Data;
+
+namespace Highway.Data
+{
+    public abstract class AdoScalarBase : QueryBase
+    {
+        protected abstract IEnumerable<IDataParameter> Parameters { get; }
+    }
+}
