+++ Highway/src/Highway.Data.EntityFramework/QueryObjects/SqlQuery.cs
@@ -3,8 +3,6 @@ using System.Collections.Generic;
 using System.Data.Common;
 using System.Data.Entity;
 using System.Data.SqlClient;
-using System.Linq;
-using System.Linq.Expressions;
 
 namespace Highway.Data
 {
