+++ Highway/src/Highway.Data.EntityFramework/QueryObjects/SqlScalar.cs
@@ -1,10 +1,6 @@
 using System;
-using System.Collections.Generic;
-using System.Data.Common;
 using System.Data.Entity;
 using System.Data.SqlClient;
-using System.Linq;
-using System.Linq.Expressions;
 
 namespace Highway.Data
 {
